Moved lib folder to global directory to remove redundancy, introduced a string for BuildInfo in build.js to reduce redundancy
