+++ bin/lib/gather-version-info.js
