+++ bin/lib/readme
